requirements.txt
# API dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
starlette>=0.27.0

# Anthropic API client
anthropic>=0.5.0

# Database
redis>=4.6.0

# Validation and settings
pydantic>=2.4.2
pydantic-settings>=2.0.3

# HTTP client
aiohttp>=3.8.5

# Data processing
PyYAML>=6.0.1

# Utilities
python-multipart>=0.0.6
ujson>=5.8.0
email-validator>=2.0.0
python-jose>=3.3.0
cryptography>=41.0.3
python-dotenv>=1.0.0